+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 ##### Unreleased
-- Nothing
+- Fixed a Chakra-based MS Edge (18-) bug that unfreeze (O_o) frozen arrays used as `WeakMap` keys
+- Fixing of the previous bug also fixes some cases of `String.dedent` in MS Edge
 
 ##### [3.27.0 - 2022.12.26](https://github.com/zloirock/core-js/releases/tag/v3.27.0)
 - [Iterator Helpers](https://github.com/tc39/proposal-iterator-helpers) proposal: